conftest.py: factor out tmpfile fixture from sage.libs.gap
A temporary file that is deleted after the test has completed will be
useful elsewhere.
